- fixed compiler warning       (Link target) ->          c:\projects\pythonscript\pythonscript\src\helpcontroller.cpp(29): warning C4706: assignment within conditional expression [C:\projects\pythonscript\PythonScript\project\PythonScript2010.vcxproj]
- code formatting

+++ PythonScript/src/ScintillaWrapper.cpp
@@ -573,7 +573,7 @@ boost::python::tuple ScintillaWrapper::getUserCharSelection()
 }
 
 
-void ScintillaWrapper::setTarget(int start, int end)
+void ScintillaWrapper::setTarget(intptr_t start, intptr_t end)
 {
 	SetTargetStart(start);
 	SetTargetEnd(end);
